+++ CHANGELOG.md
@@ -42,3 +42,5 @@
 -   Account token command.
 -   Account bot data pdf command.
 -   Account exchange code command.
+-   Account page command.
+-   Account receipts command.